Added a check to see if the bot has MESSAGE_MANAGE before trying to delete messages in Gif, Reddit and Clean commands

+++ Discord/src/commands/Gif.java
@@ -4,6 +4,7 @@ import org.jsoup.*;
 import org.jsoup.nodes.*;
 
 import backend.*;
+import net.dv8tion.jda.core.Permission;
 import net.dv8tion.jda.core.entities.*;
 import net.dv8tion.jda.core.events.message.*;
 
@@ -20,7 +21,22 @@ public class Gif {
 			channel.sendMessage("*"+event.getAuthor().getName()+"* shared a .gif of *'"+query.replace("-", " ") + "'*: " +url).queue();
 			try {
 				if(!event.getChannel().getType().equals(ChannelType.PRIVATE)){
-					event.getMessage().delete().queue();;
+					//Check if I have MESSAGE_MANAGE permission, before trying to delete
+
+					for (int i =0;i<event.getTextChannel().getMembers().size();i++) {
+						if(event.getTextChannel().getMembers().get(i).getUser().getId().equals(event.getJDA().getSelfUser().getId())){
+							//Is KakansBot
+							i=event.getTextChannel().getMembers().size()+5;
+							if(!event.getTextChannel().getMembers().get(i).hasPermission(Permission.MESSAGE_MANAGE)){
+								new Print("Cannot delete initial Gif command message in "+event.getChannel().getName()
+										+" channel in "+event.getGuild().getName()+" guild, because lack of MESSAGE_MANAGE", false);
+							}
+							else {
+								event.getMessage().delete().queue();
+							}
+						}
+					}
+
 				}
 			} catch (Exception e) {
 				// FIXME: handle exception
@@ -39,5 +55,5 @@ public class Gif {
 
 	}
 
-	
+
 }
